+++ src/main/java/edu/harvard/hul/ois/fits/Fits.java
@@ -53,7 +53,6 @@ import org.apache.commons.cli.HelpFormatter;
 import org.apache.commons.cli.Option;
 import org.apache.commons.cli.OptionGroup;
 import org.apache.commons.cli.Options;
-import org.apache.commons.cli.ParseException;
 import org.apache.commons.configuration.ConfigurationException;
 import org.apache.commons.configuration.XMLConfiguration;
 import org.apache.commons.lang.StringUtils;
@@ -81,6 +80,7 @@ public class Fits {
     private static boolean nestDirs;
     private static final XMLOutputFactory xmlOutputFactory = XMLOutputFactory.newInstance();
     private static boolean rawToolOutput = false;
+    private static boolean debug = false;
 
     private static final String FITS_CONFIG_FILE_NAME = "fits.xml";
     private static final String VERSION_PROPERTIES_FILE = "version.properties";
@@ -231,7 +231,7 @@ public class Fits {
         toolbelt = new ToolBelt(config, this);
     }
 
-    public static void main(String[] args) throws FitsException, IOException, ParseException, XMLStreamException {
+    public static void main(String[] args) throws Exception {
 
         setFitsVersionFromFile();
 
@@ -248,6 +248,7 @@ public class Fits {
         options.addOption("v", false, "print version information");
         options.addOption("f", true, "alternate fits.xml configuration file location (optional)");
         options.addOption("t", false, "include all raw tool output");
+        options.addOption("d", false, "include debug output");
         OptionGroup outputOptions = new OptionGroup();
         Option stdxml = new Option(
                 "x",
@@ -275,17 +276,10 @@ public class Fits {
             System.out.println(Fits.VERSION);
             System.exit(0);
         }
-        if (cmd.hasOption("r")) {
-            traverseDirs = true;
-        } else {
-            traverseDirs = false;
-        }
-        if (cmd.hasOption("n")) {
-            nestDirs = true;
-        } else {
-            nestDirs = false;
-        }
+        traverseDirs = cmd.hasOption("r");
+        nestDirs = cmd.hasOption("n");
         rawToolOutput = cmd.hasOption("t");
+        debug = cmd.hasOption("d");
 
         File fitsConfigFile = null;
         try {
@@ -327,9 +321,13 @@ public class Fits {
                 printHelp(options);
                 System.exit(1);
             }
-        } catch (FitsException fe) {
-            System.err.println(fe.getMessage());
-            System.exit(1);
+        } catch (Exception fe) {
+            if (debug) {
+                throw fe;
+            } else {
+                System.err.println(fe.getMessage());
+                System.exit(1);
+            }
         }
 
         System.exit(0);